port some fixes from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-2-15-branch@59195 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/Rstrptime.h
@@ -215,7 +215,7 @@ w_strptime_internal (wchar_t *rp, const wchar_t *fmt, struct tm *tm,
 		     int *poffset)
 {
     int cnt;
-    size_t val;
+    int val;
     int have_I, is_pm;
     int century, want_century;
     int have_wday, want_xday;
@@ -674,7 +674,7 @@ strptime_internal (const char *rp, const char *fmt, struct tm *tm,
 		   int *poffset)
 {
     int cnt;
-    size_t val;
+    int val;
     int have_I, is_pm;
     int century, want_century;
     int have_wday, want_xday;
